+++ src/Event/MessageEvent.php
@@ -1,0 +1,20 @@
+<?php
+
+namespace MF\ConsoleSubscriber\Event;
+
+class MessageEvent extends FormatableEvent
+{
+    /** @var string */
+    private $message;
+
+    public function __construct(string $message, string ...$args)
+    {
+        $this->message = $message;
+        $this->setArgs($args);
+    }
+
+    public function getMessage(): string
+    {
+        return $this->format($this->message);
+    }
+}