+++ src/Subscriber/ConsoleSubscriber.php
@@ -2,11 +2,15 @@
 
 namespace MF\ConsoleSubscriber\Subscriber;
 
+use MF\ConsoleSubscriber\Event\MessageEvent;
 use MF\ConsoleSubscriber\Event\NoteEvent;
 use MF\ConsoleSubscriber\Event\ProgressAdvanceEvent;
 use MF\ConsoleSubscriber\Event\ProgressFinishedEvent;
 use MF\ConsoleSubscriber\Event\ProgressStartEvent;
 use MF\ConsoleSubscriber\Event\SectionEvent;
+use MF\ConsoleSubscriber\Event\TableEvent;
+use MF\ConsoleSubscriber\Service\MessageFormatter;
+use MF\ConsoleSubscriber\Service\TableFormatter;
 use Symfony\Component\Console\Style\SymfonyStyle;
 use Symfony\Component\EventDispatcher\EventSubscriberInterface;
 
@@ -41,6 +45,8 @@ class ConsoleSubscriber implements EventSubscriberInterface
             ProgressFinishedEvent::class => ['onProgressFinished'],
             SectionEvent::class => ['onSection'],
             NoteEvent::class => ['onNote'],
+            TableEvent::class => ['onTable'],
+            MessageEvent::class => ['onMessage'],
         ];
     }
 
@@ -86,6 +92,20 @@ class ConsoleSubscriber implements EventSubscriberInterface
     {
         if ($this->io) {
             $this->io->note($event->getNote());
+        }
+    }
+
+    public function onTable(TableEvent $event)
+    {
+        if ($this->io) {
+            $this->io->table($event->getHeaders(), TableFormatter::formatRows($event->getRows()));
+        }
+    }
+
+    public function onMessage(MessageEvent $event)
+    {
+        if ($this->io) {
+            $this->io->writeln(MessageFormatter::formatMessage($event->getMessage()));
         }
     }
 }